% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hv_auth.R
\name{hv_auth}
\alias{hv_auth}
\title{Get OAuth2 token from HydroVu}
\usage{
hv_auth(
  client_id,
  client_secret,
  url = "https://www.hydrovu.com/public-api/oauth/token"
)
}
\arguments{
\item{client_id}{The client id from the API page on HydroVu}

\item{client_secret}{The client secret from the API page on HydroVu}

\item{url}{HydroVu endpoint that serves token}
}
\value{
an OAuth2 client
}
\description{
\code{hv_auth} returns a authentication token to HydroVu. Generally it should be the first
run prior to pulling more datasets.  A valid API credentials are needed on the HydroVu account.
}
\examples{
\dontrun{
token <- hv_auth('my-client-id', 'my-secret')
}
}
