% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hv_data.R
\name{hv_data}
\alias{hv_data}
\title{Get HydroVu data for a location and data range}
\usage{
hv_data(
  location_name,
  start_time,
  end_time,
  tz = "UTC",
  client,
  url = "https://www.hydrovu.com/public-api/v1/locations/"
)
}
\arguments{
\item{location_name}{Location name as provided in \code{hv_locations()}}

\item{start_time}{start date and time in ISO8601 string format:  YYYY-MM-DD HH:MM:SS}

\item{end_time}{end date and time in ISO8601 string format:  YYYY-MM-DD HH:MM:SS}

\item{tz}{time zone desired of returned data using \code{OlsonNames()}}

\item{client}{HydroVu API token as returned from \code{hv_auth()}}

\item{url}{base API url}
}
\value{
a dataframe
}
\description{
This is the main useful function of HydroVuR. Specify the Location ID and the date range to
download a dataframe of results.
}
\details{
The resulting data frame will be in "long" format with a column
for all parameters at that location. It is unknown at this time if hydrovu allows for different
time resolutions for different parameters hance casting to a wide format (one column per parameter)
could result in a sparse dataframe thus safer to export in long format

Note:  HydroVu API currently restricts API access to 1000 calls / minute and paginates ~10 records
per page.  Thus long time spans may take a while to download and/or timeout.
}
\examples{
\dontrun{
data <- hv_data("test",
                "2018-10-15 04:00", "2018-10-31 01:00", tz = "America/New_York", token)
}
}
