% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hv_locations.R
\name{hv_locations}
\alias{hv_locations}
\title{Return the list of locations for the given client from HydroVu}
\usage{
hv_locations(
  client,
  url = "https://www.hydrovu.com/public-api/v1/locations/list"
)
}
\arguments{
\item{client}{a valid OAuth2 token such as returned from \code{hv_auth()}}

\item{url}{HydroVu url that lists the locations}
}
\value{
a dataframe listing all the locations visible to the client
}
\description{
Return the list of locations for the given client from HydroVu
}
\examples{
\dontrun{
locs <- hv_locations(client)
}
}
