% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hv_names.R
\name{hv_names}
\alias{hv_names}
\title{Return the list of friendly names from HydroVu}
\usage{
hv_names(
  client,
  return = "both",
  url = "https://www.hydrovu.com/public-api/v1/sispec/friendlynames"
)
}
\arguments{
\item{client}{OAuth2 token}

\item{return}{one of "params", "units", or "both".  params and units will return a dataframe of the
respective parameters or units, both will return a list with both as separate dataframes.
Default = "both"}

\item{url}{endpoint url}
}
\value{

}
\description{
Return the list of friendly names from HydroVu
}
\examples{
\dontrun{
fNames <- hv_names(client)
}

}
